requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.3
asgiref==3.8.1
attrs==23.2.0
binutils==0.1.dev0
black==24.4.0
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
Django==4.2.11
django-allauth @ git+https://github.com/pennersr/django-allauth.git@3582dea795903766831c59400c77eb0c90743001
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-ipware==6.0.5
django-phonenumber-field==7.3.0
django-rest-knox @ git+https://github.com/jazzband/django-rest-knox.git@09403255a069c2416a5dfb889f80342a2d7b2646
django-twilio==0.14.3.3
djangorestframework==3.15.1
factory-boy==3.3.0
Faker==24.4.0
filelock==3.13.3
flake8==7.0.0
frozenlist==1.4.1
glibc==0.6.1
idna==3.4
inflection==0.5.1
mccabe==0.7.0
multidict==6.0.5
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.34
phonenumberslite==8.13.34
pillow==10.3.0
platformdirs==4.2.0
psycopg==3.1.18
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2022.12
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ipware==2.0.3
python-pip==1.1.1
python3-openid==3.2.0
qrcode==7.4.2
requests==2.28.1
requests-oauthlib==2.0.0
semantic-version==2.10.0
setuptools-rust==1.9.0
six==1.16.0
sqlparse==0.4.4
toposort==1.10
twilio==9.0.4
tzdata==2024.1
urllib3==1.26.12
virtualenv==20.25.1
yarl==1.9.4
